+++ pkgs/intl4x/lib/src/display_names/display_names_impl.dart
@@ -9,7 +9,8 @@ import '../ecma/ecma_policy.dart';
 import '../locale/locale.dart';
 import '../options.dart';
 import '../utils.dart';
-import 'display_names_4x.dart';
+import 'display_names_4x.dart'
+    if (dart.library.js) 'display_names_stub_4x.dart';
 import 'display_names_options.dart';
 import 'display_names_stub.dart' if (dart.library.js) 'display_names_ecma.dart';